集成DB, 完成后端api权限认证, 采用mybatis二级缓存，用rediss作为缓存中间件(未提供更新db权限时并更新缓存的操作，待更…

+++ main/src/main/java/com/eugene/sumarry/permission/dao/UserDao.java
@@ -3,7 +3,7 @@ package com.eugene.sumarry.permission.dao;
 import com.eugene.sumarry.permission.model.User;
 import org.apache.ibatis.annotations.Param;
 
-public interface UserDao extends BaseDao<User, Integer> {
+public interface UserDao extends BaseDao<User, Long> {
 
     User checkLoginInfo(@Param("userName") String userName, @Param("password") String password);
 }
